# core/vm: Refactor jump table to merge dynamic gas cost with memory expansion cost calculations