--- a/core/vm/memory_table.go
+++ b/core/vm/memory_table.go
@@ -67,7 +67,3 @@ func memoryStaticCall(stack *Stack) (uint64, bool) {
 	}
 	return y, false
 }
-
-func memoryLog(stack *Stack) (uint64, bool) {
-	return calcMemSize64(stack.Back(0), stack.Back(1))
-}