This was discussed in https://github.com/ethereum/go-ethereum/pull/24048#discussion_r761849601

Basically it merges `operation.dynamicGas` call with `operation.memorySize` call into single `dynamicGas` call returning both gas required and new memory size (or 0 if memory size doesn't change).

It allows to simplify logic in interpreter loop and also to get rid of init-time validation to check that `operation.memory != nil` iff `operation.dynamicGas != nil` (Because it's now expected from `dynamicGas` function author to correctly cover both costs, `operation.memory` member is removed)