@holiman 2021-12-16
As for usual geth-style, named returns are ok, but silent/implicit returns are no-no