@gumb0 2021-12-16
Not much changes in my benchmarks:
```
name                                  old time/op  new time/op  delta
main/blake2b_shifts/2805nulls         12.3ms ±13%  12.7ms ±12%    ~     (p=0.529 n=20+20)
main/blake2b_shifts/5610nulls         26.6ms ± 1%  26.5ms ± 1%  -0.57%  (p=0.000 n=19+18)
main/blake2b_shifts/8415nulls         39.6ms ± 2%  39.2ms ± 2%  -0.87%  (p=0.003 n=19+18)
main/weierstrudel/0                    454µs ± 0%   458µs ± 0%  +0.82%  (p=0.000 n=19+19)
main/weierstrudel/1                    918µs ± 0%   930µs ± 1%  +1.27%  (p=0.000 n=20+20)
main/weierstrudel/3                   1.44ms ± 0%  1.46ms ± 1%  +1.26%  (p=0.000 n=19+18)
main/weierstrudel/9                   2.97ms ± 1%  3.00ms ± 1%  +0.95%  (p=0.000 n=18+19)
main/weierstrudel/14                  4.24ms ± 1%  4.28ms ± 1%  +0.94%  (p=0.000 n=18+19)
main/sha1_divs/empty                   181µs ± 1%   181µs ± 1%  +0.41%  (p=0.001 n=19+20)
main/sha1_divs/1351                   3.54ms ± 1%  3.54ms ± 1%    ~     (p=0.945 n=19+20)
main/sha1_divs/2737                   6.92ms ± 2%  6.87ms ± 1%  -0.82%  (p=0.000 n=19+19)
main/sha1_divs/5311                   13.4ms ± 3%  13.5ms ± 2%  +0.85%  (p=0.004 n=18+19)
main/sha1_shifts/empty                 142µs ± 1%   146µs ± 1%  +2.88%  (p=0.000 n=18+20)
main/sha1_shifts/1351                 2.82ms ± 1%  2.89ms ± 1%  +2.28%  (p=0.000 n=18+19)
main/sha1_shifts/2737                 5.49ms ± 1%  5.58ms ± 1%  +1.66%  (p=0.000 n=18+20)
main/sha1_shifts/5311                 10.8ms ± 1%  11.0ms ± 1%  +1.85%  (p=0.000 n=18+17)
main/blake2b_huff/empty                100µs ± 0%   100µs ± 1%    ~     (p=0.746 n=17+20)
main/blake2b_huff/2805nulls           1.89ms ± 1%  1.89ms ± 1%    ~     (p=0.779 n=20+20)
main/blake2b_huff/5610nulls           3.63ms ± 1%  3.63ms ± 1%    ~     (p=0.443 n=18+20)
main/blake2b_huff/8415nulls           5.31ms ± 1%  5.30ms ± 1%    ~     (p=0.141 n=20+19)
micro/JUMPDEST_n0/empty               2.21ms ± 1%  2.19ms ± 1%  -1.24%  (p=0.000 n=20+18)
micro/jump_around/empty                136µs ± 1%   136µs ± 1%    ~     (p=0.067 n=20+18)
micro/memory_grow_mstore/nogrow        434µs ± 1%   434µs ± 1%    ~     (p=0.365 n=19+20)
micro/memory_grow_mstore/by1           458µs ± 0%   459µs ± 1%    ~     (p=0.322 n=19+20)
micro/memory_grow_mstore/by16          544µs ± 1%   544µs ± 1%    ~     (p=0.808 n=18+18)
micro/memory_grow_mstore/by32          652µs ± 0%   652µs ± 1%    ~     (p=0.900 n=17+19)
micro/signextend/zero                  304µs ± 1%   302µs ± 1%  -0.57%  (p=0.008 n=19+19)
micro/signextend/one                   321µs ± 5%   312µs ± 1%  -2.78%  (p=0.000 n=20+19)
micro/loop_with_many_jumpdests/empty  1.83µs ± 4%  1.85µs ± 2%    ~     (p=0.173 n=18+20)
micro/memory_grow_mload/nogrow         202µs ± 1%   205µs ± 1%  +1.03%  (p=0.000 n=20+19)
micro/memory_grow_mload/by1            218µs ± 1%   221µs ± 1%  +1.38%  (p=0.000 n=19+20)
micro/memory_grow_mload/by16           303µs ± 1%   305µs ± 0%  +0.78%  (p=0.000 n=18+19)
micro/memory_grow_mload/by32           398µs ± 1%   400µs ± 1%  +0.47%  (p=0.000 n=19+20)
[Geo mean]                            1.07ms       1.08ms       +0.47%
```