@holiman 2021-12-16
As it is, I'm a bit torn on this. It is a pretty deep change, which needs to be very thorougly reviewed, but the benchmarks do not really support the change. 
However, I think you may have missed an opportunity for optimization: previously, the code used various `makeMemoryCopier` bla bla, to build a function that we shoved into a map. Since we were going to do the call into it via an indirecton in `operation`, there was no way it was going to be inlined, so it didn't matter. 

Now that you placed _into_ dynamicGas, it might be possible to make the call to the memory calculation become inlined. But it might mean that instead of calling `makeMemoryCoper`, you would instead define each such function. And that might improve the benchmark. Just a thought/theory.