@gumb0 2021-12-21
> However, I think you may have missed an opportunity for optimization: previously, the code used various makeMemoryCopier bla bla, to build a function that we shoved into a map. Since we were going to do the call into it via an indirecton in operation, there was no way it was going to be inlined, so it didn't matter.
>
>Now that you placed into dynamicGas, it might be possible to make the call to the memory calculation become inlined. But it might mean that instead of calling makeMemoryCoper, you would instead define each such function. And that might improve the benchmark. Just a thought/theory.

In this refactoring I keep the memory gas calculation general, by having low-level `memoryGasCostAndSize` function which takes as a paramter a function returning memory size for particular opcode. So there's still indirection (even if particular `memoryCopierGas` is unrolled) and unless we copypaste this `memoryGasCostAndSize`  into memory expanding opcodes' gas calculations, I doubt it will get to inlining. And it would affect only this small subset of opcodes that expand memory. But I can try.